added player stats is now keyed by uuid, rather than player interface

+++ src/main/java/com/runicrealms/runicitems/config/ItemLoader.java
@@ -77,7 +77,7 @@ public class ItemLoader {
             Set<String> sectionKeys = section.getSection("stats").getKeys();
             for (PlayerStatEnum templateStatType : templateStats.keySet()) {
                 if (sectionKeys.contains(templateStatType.getName())) { // Item has stat and has already been rolled (stored in database)
-                    stats.put(templateStatType, new RunicItemStat(templateStats.get(templateStatType), section.get("stats." + templateStatType.getName(), Float.class)));
+                    stats.put(templateStatType, new RunicItemStat(templateStats.get(templateStatType), section.get("stats." + templateStatType.getName(), Double.class)));
                 } else { // Item has recently been added this stat and the roll does not exist in database, so make a new roll
                     stats.put(templateStatType, new RunicItemStat(templateStats.get(templateStatType)));
                 }
